# Fix(GraphQL): Refactor Mutation Rewriter for Add and Update Mutations