--- a/ee/acl/acl_test.go
+++ b/ee/acl/acl_test.go
@@ -230,9 +230,8 @@ func TestCreateAndDeleteUsers(t *testing.T) {
 	token := testutil.GrootHttpLogin(adminEndpoint)
 	resp := createUser(t, token, userid, userpassword)
 	require.Equal(t, x.GqlErrorList{{
-		Message: "couldn't rewrite query for mutation addUser because id alice already exists" +
-			" for type User",
-		Path: []interface{}{"addUser"},
+		Message: "couldn't rewrite mutation addUser because failed to rewrite mutation payload because id" +
+			" alice already exists for type User",
 	}}, resp.Errors)
 	checkUserCount(t, resp.Data, 0)