--- a/ee/acl/acl_test.go
+++ b/ee/acl/acl_test.go
@@ -229,10 +229,9 @@ func TestCreateAndDeleteUsers(t *testing.T) {
 	// adding the user again should fail
 	token := testutil.GrootHttpLogin(adminEndpoint)
 	resp := createUser(t, token, userid, userpassword)
-	require.Equal(t, x.GqlErrorList{{
-		Message: "couldn't rewrite mutation addUser because failed to rewrite mutation payload because id" +
-			" alice already exists for type User",
-	}}, resp.Errors)
+	require.Equal(t, 1, len(resp.Errors))
+	require.Equal(t, "couldn't rewrite mutation addUser because failed to rewrite mutation payload because id"+
+		" alice already exists for type User", resp.Errors[0].Message)
 	checkUserCount(t, resp.Data, 0)
 
 	// delete the user