--- a/graphql/e2e/common/common.go
+++ b/graphql/e2e/common/common.go
@@ -774,6 +774,7 @@ func RunAll(t *testing.T) {
 	t.Run("three level double XID mutation", threeLevelDoubleXID)
 	t.Run("two levels linked to one XID", twoLevelsLinkedToXID)
 	t.Run("cyclically linked mutation", cyclicMutation)
+	t.Run("parallel mutations", parallelMutations)
 
 	// error tests
 	t.Run("graphql completion on", graphQLCompletionOn)