--- a/graphql/resolve/mutation.go
+++ b/graphql/resolve/mutation.go
@@ -20,6 +20,7 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
+	"fmt"
 	"sort"
 	"strconv"
 
@@ -30,6 +31,7 @@ import (
 	"github.com/dgraph-io/dgraph/graphql/schema"
 	"github.com/dgraph-io/dgraph/x"
 	"github.com/golang/glog"
+	"github.com/pkg/errors"
 	otrace "go.opencensus.io/trace"
 )
 
@@ -196,6 +198,7 @@ func getNumUids(m schema.Mutation, a map[string]string, r map[string]interface{}
 func (mr *dgraphResolver) rewriteAndExecute(ctx context.Context,
 	mutation schema.Mutation) (*Resolved, bool) {
 	var mutResp, qryResp *dgoapi.Response
+	req := &dgoapi.Request{}
 	commit := false
 
 	defer func() {
@@ -238,7 +241,90 @@ func (mr *dgraphResolver) rewriteAndExecute(ctx context.Context,
 		}
 	}
 
-	upserts, err := mr.mutationRewriter.Rewrite(ctx, mutation)
+	var upserts []*UpsertMutation
+	var err error
+	var frags [][]*mutationFragment
+	switch mr.mutationRewriter.(type) {
+	case *AddRewriter:
+		varGen := NewVariableGenerator()
+		xidMd := newXidMetadata()
+		var queries []*gql.GraphQuery
+		queries, err = NewRewrite(ctx, mutation, varGen, xidMd)
+		if err != nil {
+			return emptyResult(schema.GQLWrapf(err, "couldn't rewrite mutation %s", mutation.Name())),
+				resolverFailed
+		}
+		// Execute queries and parse its result into a map
+		qry := dgraph.AsString(queries)
+		req.Query = qry
+		var mutResp *dgoapi.Response
+
+		// The query will be empty in case there is no reference XID / UID in the mutation.
+		// Don't execute the query in those cases.
+		if req.Query != "" {
+			mutResp, err = mr.executor.Execute(ctx, req, nil)
+		}
+		if err != nil {
+			gqlErr := schema.GQLWrapLocationf(
+				err, mutation.Location(), "mutation %s failed", mutation.Name())
+			return emptyResult(gqlErr), resolverFailed
+		}
+
+		// Parse the result of query.
+		// mutResp.Json will contain response to the query.
+		// The response is parsed to existenceQueriesResult
+		// Example Response:
+		// {
+		// 	Project1 :
+		//		[
+		//			{
+		//				"uid" : "0x123"
+		// 			}
+		//		],
+		//	Column2 :
+		//		[
+		//			{
+		//				"uid": "0x234"
+		// 			}
+		//		]
+		// }
+		queryResultMap := make(map[string][]map[string]string)
+		if mutResp != nil {
+			err = json.Unmarshal(mutResp.Json, &queryResultMap)
+		}
+		if err != nil {
+			gqlErr := schema.GQLWrapLocationf(
+				err, mutation.Location(), "mutation %s failed", mutation.Name())
+			return emptyResult(gqlErr), resolverFailed
+		}
+
+		// The above response is parsed into map[string]string as follows:
+		// {
+		// 		"Project1" : "0x123",
+		// 		"Column2" : "0x234"
+		// }
+		qNameToUID := make(map[string]string)
+		for key, result := range queryResultMap {
+			if len(result) == 1 {
+				// Found exactly one UID / XID corresponding to given condition
+				qNameToUID[key] = result[0]["uid"]
+			} else if len(result) > 1 {
+				// Found multiple UIDs for query. This should ideally not happen.
+				// This indicates that there are multiple nodes with same XIDs / UIDs. Throw an error.
+				err = errors.New(fmt.Sprintf("Found multiple nodes with ID: %s", result[0]["uid"]))
+				gqlErr := schema.GQLWrapLocationf(
+					err, mutation.Location(), "mutation %s failed", mutation.Name())
+				return emptyResult(gqlErr), resolverFailed
+			}
+		}
+
+		// Create mutations and new nodes depending on result of queries.
+		upserts, frags, err = NewCreateMutations(ctx, mutation, varGen, xidMd, qNameToUID)
+
+	default:
+		upserts, err = mr.mutationRewriter.Rewrite(ctx, mutation)
+	}
+
 	if err != nil {
 		return emptyResult(schema.GQLWrapf(err, "couldn't rewrite mutation %s", mutation.Name())),
 			resolverFailed
@@ -278,7 +364,6 @@ func (mr *dgraphResolver) rewriteAndExecute(ctx context.Context,
 	}
 
 	result := make(map[string]interface{})
-	req := &dgoapi.Request{}
 	newNodes := make(map[string]schema.Type)
 
 	mutationTimer := newtimer(ctx, &dgraphMutationDuration.OffsetDuration)
@@ -313,8 +398,16 @@ func (mr *dgraphResolver) rewriteAndExecute(ctx context.Context,
 		return emptyResult(schema.GQLWrapf(authErr, "mutation failed")), resolverFailed
 	}
 
-	dgQuery, err := mr.mutationRewriter.FromMutationResult(ctx, mutation, mutResp.GetUids(), result)
-	queryErrs = schema.AppendGQLErrs(queryErrs, schema.GQLWrapf(err,
+	var errs error
+	var dgQuery []*gql.GraphQuery
+	switch mr.mutationRewriter.(type) {
+	case *AddRewriter:
+		dgQuery, err = NewFromMutationResult(ctx, mutation, mutResp.GetUids(), result, frags)
+	default:
+		dgQuery, err = mr.mutationRewriter.FromMutationResult(ctx, mutation, mutResp.GetUids(), result)
+
+	}
+	errs = schema.AppendGQLErrs(errs, schema.GQLWrapf(err,
 		"couldn't rewrite query for mutation %s", mutation.Name()))
 	if dgQuery == nil && err != nil {
 		return emptyResult(queryErrs), resolverFailed