… graphql/e2e/directives/schema_response.json
… graphql/e2e/normal/schema_response.json
… graphql/resolve/add_mutation_test.yaml
… graphql/resolve/auth_add_test.yaml
… graphql/resolve/auth_update_test.yaml
… graphql/resolve/extensions_test.go
… graphql/resolve/mutation_rewriter.go
… graphql/resolve/mutation_test.go
… graphql/resolve/update_mutation_test.yaml