Motivation:
Recently, two separate PRs for add and update mutation rewriting refactor were merged to graphql_mutation_refactor branch. This PR is about merging all those changes to maser.

Related PRs:
1. https://github.com/dgraph-io/dgraph/pull/7383
2. https://github.com/dgraph-io/dgraph/pull/7257

Benchmarking Results for graphql/resolve/mutation_test.go:3LevelDeep10000
Master branch: (Before refactoring)
pkg: github.com/dgraph-io/dgraph/graphql/resolve
Benchmark3LevelDeep10000-8   	       1	22271327591 ns/op
PASS
ok  	github.com/dgraph-io/dgraph/graphql/resolve	22.418s

Current branch: (After refactoring)
pkg: github.com/dgraph-io/dgraph/graphql/resolve
Benchmark3LevelDeep10000-8   	      26	  45202006 ns/op
PASS
ok  	github.com/dgraph-io/dgraph/graphql/resolve	2.708s

Fixes GRAPHQL-962 and GRAPHQL-872
Also fixes GRAPHQL-623, GRAPHQL-803 .
<!--
Your title must be in the following format: topic(Area): Feature
Topic must be one of build|ci|docs|feat|fix|perf|refactor|chore|test

Sample Titles:
feat(Enterprise): Backups can now get credentials from IAM
fix(Query): Skipping floats that cannot be Marshalled in JSON
perf: [Breaking] json encoding is now 35% faster if SIMD is present
chore: all chores/tests will be excluded from the CHANGELOG

Please add a description with these things:
1. A good description explaining the problem and what you changed.
2. If it fixes any GitHub issues, say "Fixes #GitHubIssue".
3. If it corresponds to a Jira issue, say "Fixes DGRAPH-###".
4. If this is a breaking change, please put "[Breaking]" in the title. In the description, please put a note with exactly who these changes are breaking for.
-->

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/dgraph/7409)
<!-- Reviewable:end -->